refactor out filter_merge references
