+++ daemons/Makefile
@@ -1,2 +1,2 @@
 build clean clobber deploy:
-	$(MAKE) -C filter_merge $@
+	$(MAKE) -C driver $@
